+++ tests/smoke_tests/__init__.py
@@ -1,0 +1,2 @@
+"""For smoke tests import."""
+__all__ = ['util']
