+++ tests/smoke_tests/test_smoke.py
@@ -44,6 +44,27 @@ import uuid
 import colorama
 import jinja2
 import pytest
+from smoke_tests.util import _get_cluster_name
+from smoke_tests.util import (
+    _get_cmd_wait_until_cluster_status_contains_wildcard)
+from smoke_tests.util import _GET_JOB_QUEUE
+from smoke_tests.util import _get_timeout
+from smoke_tests.util import _JOB_WAIT_NOT_RUNNING
+from smoke_tests.util import _VALIDATE_LAUNCH_OUTPUT
+from smoke_tests.util import _WAIT_UNTIL_CLUSTER_IS_NOT_FOUND
+from smoke_tests.util import _WAIT_UNTIL_CLUSTER_STATUS_CONTAINS
+from smoke_tests.util import _WAIT_UNTIL_JOB_STATUS_CONTAINS_MATCHING_JOB_ID
+from smoke_tests.util import (
+    _WAIT_UNTIL_JOB_STATUS_CONTAINS_WITHOUT_MATCHING_JOB)
+from smoke_tests.util import (
+    _WAIT_UNTIL_MANAGED_JOB_STATUS_CONTAINS_MATCHING_JOB_NAME)
+from smoke_tests.util import FLUIDSTACK_TYPE
+from smoke_tests.util import LAMBDA_TYPE
+from smoke_tests.util import run_one_test
+from smoke_tests.util import SCP_GPU_V100
+from smoke_tests.util import SCP_TYPE
+from smoke_tests.util import STORAGE_SETUP_COMMANDS
+from smoke_tests.util import Test
 
 import sky
 from sky import global_user_state
@@ -67,1020 +88,6 @@ from sky.status_lib import ClusterStatus
 from sky.utils import common_utils
 from sky.utils import resources_utils
 from sky.utils import subprocess_utils
-
-# To avoid the second smoke test reusing the cluster launched in the first
-# smoke test. Also required for test_managed_jobs_recovery to make sure the
-# manual termination with aws ec2 does not accidentally terminate other clusters
-# for for the different managed jobs launch with the same job name but a
-# different job id.
-test_id = str(uuid.uuid4())[-2:]
-
-LAMBDA_TYPE = '--cloud lambda --gpus A10'
-FLUIDSTACK_TYPE = '--cloud fluidstack --gpus RTXA4000'
-
-SCP_TYPE = '--cloud scp'
-SCP_GPU_V100 = '--gpus V100-32GB'
-
-STORAGE_SETUP_COMMANDS = [
-    'touch ~/tmpfile', 'mkdir -p ~/tmp-workdir',
-    'touch ~/tmp-workdir/tmp\ file', 'touch ~/tmp-workdir/tmp\ file2',
-    'touch ~/tmp-workdir/foo',
-    '[ ! -e ~/tmp-workdir/circle-link ] && ln -s ~/tmp-workdir/ ~/tmp-workdir/circle-link || true',
-    'touch ~/.ssh/id_rsa.pub'
-]
-
-# Get the job queue, and print it once on its own, then print it again to
-# use with grep by the caller.
-_GET_JOB_QUEUE = 's=$(sky jobs queue); echo "$s"; echo "$s"'
-# Wait for a job to be not in RUNNING state. Used to check for RECOVERING.
-_JOB_WAIT_NOT_RUNNING = (
-    's=$(sky jobs queue);'
-    'until ! echo "$s" | grep "{job_name}" | grep "RUNNING"; do '
-    'sleep 10; s=$(sky jobs queue);'
-    'echo "Waiting for job to stop RUNNING"; echo "$s"; done')
-
-# Cluster functions
-_ALL_JOB_STATUSES = "|".join([status.value for status in JobStatus])
-_ALL_CLUSTER_STATUSES = "|".join([status.value for status in ClusterStatus])
-_ALL_MANAGED_JOB_STATUSES = "|".join(
-    [status.value for status in ManagedJobStatus])
-
-_WAIT_UNTIL_CLUSTER_STATUS_CONTAINS = (
-    # A while loop to wait until the cluster status
-    # becomes certain status, with timeout.
-    'start_time=$SECONDS; '
-    'while true; do '
-    'if (( $SECONDS - $start_time > {timeout} )); then '
-    '  echo "Timeout after {timeout} seconds waiting for cluster status \'{cluster_status}\'"; exit 1; '
-    'fi; '
-    'current_status=$(sky status {cluster_name} --refresh | '
-    'awk "/^{cluster_name}/ '
-    '{{for (i=1; i<=NF; i++) if (\$i ~ /^(' + _ALL_CLUSTER_STATUSES +
-    ')$/) print \$i}}"); '
-    'if [[ "$current_status" =~ {cluster_status} ]]; '
-    'then echo "Target cluster status {cluster_status} reached."; break; fi; '
-    'echo "Waiting for cluster status to become {cluster_status}, current status: $current_status"; '
-    'sleep 10; '
-    'done')
-
-
-def _get_cmd_wait_until_cluster_status_contains_wildcard(
-        cluster_name_wildcard: str, cluster_status: str, timeout: int):
-    wait_cmd = _WAIT_UNTIL_CLUSTER_STATUS_CONTAINS.replace(
-        'sky status {cluster_name}',
-        'sky status "{cluster_name}"').replace('awk "/^{cluster_name}/',
-                                               'awk "/^{cluster_name_awk}/')
-    return wait_cmd.format(cluster_name=cluster_name_wildcard,
-                           cluster_name_awk=cluster_name_wildcard.replace(
-                               '*', '.*'),
-                           cluster_status=cluster_status,
-                           timeout=timeout)
-
-
-_WAIT_UNTIL_CLUSTER_IS_NOT_FOUND = (
-    # A while loop to wait until the cluster is not found or timeout
-    'start_time=$SECONDS; '
-    'while true; do '
-    'if (( $SECONDS - $start_time > {timeout} )); then '
-    '  echo "Timeout after {timeout} seconds waiting for cluster to be removed"; exit 1; '
-    'fi; '
-    'if sky status -r {cluster_name}; sky status {cluster_name} | grep "{cluster_name} not found"; then '
-    '  echo "Cluster {cluster_name} successfully removed."; break; '
-    'fi; '
-    'echo "Waiting for cluster {name} to be removed..."; '
-    'sleep 10; '
-    'done')
-
-_WAIT_UNTIL_JOB_STATUS_CONTAINS_MATCHING_JOB_ID = (
-    # A while loop to wait until the job status
-    # contains certain status, with timeout.
-    'start_time=$SECONDS; '
-    'while true; do '
-    'if (( $SECONDS - $start_time > {timeout} )); then '
-    '  echo "Timeout after {timeout} seconds waiting for job status \'{job_status}\'"; exit 1; '
-    'fi; '
-    'current_status=$(sky queue {cluster_name} | '
-    'awk "\\$1 == \\"{job_id}\\" '
-    '{{for (i=1; i<=NF; i++) if (\$i ~ /^(' + _ALL_JOB_STATUSES +
-    ')$/) print \$i}}"); '
-    'found=0; '  # Initialize found variable outside the loop
-    'while read -r line; do '  # Read line by line
-    '  if [[ "$line" =~ {job_status} ]]; then '  # Check each line
-    '    echo "Target job status {job_status} reached."; '
-    '    found=1; '
-    '    break; '  # Break inner loop
-    '  fi; '
-    'done <<< "$current_status"; '
-    'if [ "$found" -eq 1 ]; then break; fi; '  # Break outer loop if match found
-    'echo "Waiting for job status to contains {job_status}, current status: $current_status"; '
-    'sleep 10; '
-    'done')
-
-_WAIT_UNTIL_JOB_STATUS_CONTAINS_WITHOUT_MATCHING_JOB = _WAIT_UNTIL_JOB_STATUS_CONTAINS_MATCHING_JOB_ID.replace(
-    'awk "\\$1 == \\"{job_id}\\"', 'awk "')
-
-_WAIT_UNTIL_JOB_STATUS_CONTAINS_MATCHING_JOB_NAME = _WAIT_UNTIL_JOB_STATUS_CONTAINS_MATCHING_JOB_ID.replace(
-    'awk "\\$1 == \\"{job_id}\\"', 'awk "\\$2 == \\"{job_name}\\"')
-
-# Managed job functions
-
-_WAIT_UNTIL_MANAGED_JOB_STATUS_CONTAINS_MATCHING_JOB_NAME = _WAIT_UNTIL_JOB_STATUS_CONTAINS_MATCHING_JOB_NAME.replace(
-    'sky queue {cluster_name}', 'sky jobs queue').replace(
-        'awk "\\$2 == \\"{job_name}\\"',
-        'awk "\\$2 == \\"{job_name}\\" || \\$3 == \\"{job_name}\\"').replace(
-            _ALL_JOB_STATUSES, _ALL_MANAGED_JOB_STATUSES)
-
-# After the timeout, the cluster will stop if autostop is set, and our check
-# should be more than the timeout. To address this, we extend the timeout by
-# _BUMP_UP_SECONDS before exiting.
-_BUMP_UP_SECONDS = 35
-
-DEFAULT_CMD_TIMEOUT = 15 * 60
-
-
-class Test(NamedTuple):
-    name: str
-    # Each command is executed serially.  If any failed, the remaining commands
-    # are not run and the test is treated as failed.
-    commands: List[str]
-    teardown: Optional[str] = None
-    # Timeout for each command in seconds.
-    timeout: int = DEFAULT_CMD_TIMEOUT
-    # Environment variables to set for each command.
-    env: Dict[str, str] = None
-
-    def echo(self, message: str):
-        # pytest's xdist plugin captures stdout; print to stderr so that the
-        # logs are streaming while the tests are running.
-        prefix = f'[{self.name}]'
-        message = f'{prefix} {message}'
-        message = message.replace('\n', f'\n{prefix} ')
-        print(message, file=sys.stderr, flush=True)
-
-
-def _get_timeout(generic_cloud: str,
-                 override_timeout: int = DEFAULT_CMD_TIMEOUT):
-    timeouts = {'fluidstack': 60 * 60}  # file_mounts
-    return timeouts.get(generic_cloud, override_timeout)
-
-
-def _get_cluster_name() -> str:
-    """Returns a user-unique cluster name for each test_<name>().
-
-    Must be called from each test_<name>().
-    """
-    caller_func_name = inspect.stack()[1][3]
-    test_name = caller_func_name.replace('_', '-').replace('test-', 't-')
-    test_name = common_utils.make_cluster_name_on_cloud(test_name,
-                                                        24,
-                                                        add_user_hash=False)
-    return f'{test_name}-{test_id}'
-
-
-def _terminate_gcp_replica(name: str, zone: str, replica_id: int) -> str:
-    cluster_name = serve.generate_replica_cluster_name(name, replica_id)
-    query_cmd = (f'gcloud compute instances list --filter='
-                 f'"(labels.ray-cluster-name:{cluster_name})" '
-                 f'--zones={zone} --format="value(name)"')
-    return (f'gcloud compute instances delete --zone={zone}'
-            f' --quiet $({query_cmd})')
-
-
-def run_one_test(test: Test) -> Tuple[int, str, str]:
-    # Fail fast if `sky` CLI somehow errors out.
-    subprocess.run(['sky', 'status'], stdout=subprocess.DEVNULL, check=True)
-    log_file = tempfile.NamedTemporaryFile('a',
-                                           prefix=f'{test.name}-',
-                                           suffix='.log',
-                                           delete=False)
-    test.echo(f'Test started. Log: less {log_file.name}')
-    env_dict = os.environ.copy()
-    if test.env:
-        env_dict.update(test.env)
-    for command in test.commands:
-        log_file.write(f'+ {command}\n')
-        log_file.flush()
-        proc = subprocess.Popen(
-            command,
-            stdout=log_file,
-            stderr=subprocess.STDOUT,
-            shell=True,
-            executable='/bin/bash',
-            env=env_dict,
-        )
-        try:
-            proc.wait(timeout=test.timeout)
-        except subprocess.TimeoutExpired as e:
-            log_file.flush()
-            test.echo(f'Timeout after {test.timeout} seconds.')
-            test.echo(str(e))
-            log_file.write(f'Timeout after {test.timeout} seconds.\n')
-            log_file.flush()
-            # Kill the current process.
-            proc.terminate()
-            proc.returncode = 1  # None if we don't set it.
-            break
-
-        if proc.returncode:
-            break
-
-    style = colorama.Style
-    fore = colorama.Fore
-    outcome = (f'{fore.RED}Failed{style.RESET_ALL}'
-               if proc.returncode else f'{fore.GREEN}Passed{style.RESET_ALL}')
-    reason = f'\nReason: {command}' if proc.returncode else ''
-    msg = (f'{outcome}.'
-           f'{reason}'
-           f'\nLog: less {log_file.name}\n')
-    test.echo(msg)
-    log_file.write(msg)
-    if (proc.returncode == 0 or
-            pytest.terminate_on_failure) and test.teardown is not None:
-        subprocess_utils.run(
-            test.teardown,
-            stdout=log_file,
-            stderr=subprocess.STDOUT,
-            timeout=10 * 60,  # 10 mins
-            shell=True,
-        )
-
-    if proc.returncode:
-        raise Exception(f'test failed: less {log_file.name}')
-
-
-def get_aws_region_for_quota_failover() -> Optional[str]:
-    candidate_regions = AWS.regions_with_offering(instance_type='p3.16xlarge',
-                                                  accelerators=None,
-                                                  use_spot=True,
-                                                  region=None,
-                                                  zone=None)
-    original_resources = sky.Resources(cloud=sky.AWS(),
-                                       instance_type='p3.16xlarge',
-                                       use_spot=True)
-
-    # Filter the regions with proxy command in ~/.sky/config.yaml.
-    filtered_regions = original_resources.get_valid_regions_for_launchable()
-    candidate_regions = [
-        region for region in candidate_regions
-        if region.name in filtered_regions
-    ]
-
-    for region in candidate_regions:
-        resources = original_resources.copy(region=region.name)
-        if not AWS.check_quota_available(resources):
-            return region.name
-
-    return None
-
-
-def get_gcp_region_for_quota_failover() -> Optional[str]:
-
-    candidate_regions = GCP.regions_with_offering(instance_type=None,
-                                                  accelerators={'A100-80GB': 1},
-                                                  use_spot=True,
-                                                  region=None,
-                                                  zone=None)
-
-    original_resources = sky.Resources(cloud=sky.GCP(),
-                                       instance_type='a2-ultragpu-1g',
-                                       accelerators={'A100-80GB': 1},
-                                       use_spot=True)
-
-    # Filter the regions with proxy command in ~/.sky/config.yaml.
-    filtered_regions = original_resources.get_valid_regions_for_launchable()
-    candidate_regions = [
-        region for region in candidate_regions
-        if region.name in filtered_regions
-    ]
-
-    for region in candidate_regions:
-        if not GCP.check_quota_available(
-                original_resources.copy(region=region.name)):
-            return region.name
-
-    return None
-
-
-# ---------- Dry run: 2 Tasks in a chain. ----------
-@pytest.mark.no_fluidstack  #requires GCP and AWS set up
-def test_example_app():
-    test = Test(
-        'example_app',
-        ['python examples/example_app.py'],
-    )
-    run_one_test(test)
-
-
-_VALIDATE_LAUNCH_OUTPUT = (
-    # Validate the output of the job submission:
-    # ⚙️ Launching on Kubernetes.
-    #   Pod is up.
-    # ✓ Cluster launched: test. View logs at: ~/sky_logs/sky-2024-10-07-19-44-18-177288/provision.log
-    # ⚙️ Running setup on 1 pod.
-    # running setup
-    # ✓ Setup completed.
-    # ⚙️ Job submitted, ID: 1.
-    # ├── Waiting for task resources on 1 node.
-    # └── Job started. Streaming logs... (Ctrl-C to exit log streaming; job will not be killed)
-    # (min, pid=1277) # conda environments:
-    # (min, pid=1277) #
-    # (min, pid=1277) base                  *  /opt/conda
-    # (min, pid=1277)
-    # (min, pid=1277) task run finish
-    # ✓ Job finished (status: SUCCEEDED).
-    #
-    # Job ID: 1
-    # 📋 Useful Commands
-    # ├── To cancel the job:          sky cancel test 1
-    # ├── To stream job logs:         sky logs test 1
-    # └── To view job queue:          sky queue test
-    #
-    # Cluster name: test
-    # ├── To log into the head VM:    ssh test
-    # ├── To submit a job:            sky exec test yaml_file
-    # ├── To stop the cluster:        sky stop test
-    # └── To teardown the cluster:    sky down test
-    'echo "$s" && echo "==Validating launching==" && '
-    'echo "$s" | grep -A 1 "Launching on" | grep "is up." && '
-    'echo "$s" && echo "==Validating setup output==" && '
-    'echo "$s" | grep -A 1 "Running setup on" | grep "running setup" && '
-    'echo "==Validating running output hints==" && echo "$s" | '
-    'grep -A 1 "Job submitted, ID:" | '
-    'grep "Waiting for task resources on " && '
-    'echo "==Validating task output starting==" && echo "$s" | '
-    'grep -A 1 "Job started. Streaming logs..." | grep "(min, pid=" && '
-    'echo "==Validating task output ending==" && '
-    'echo "$s" | grep -A 1 "task run finish" | '
-    'grep "Job finished (status: SUCCEEDED)" && '
-    'echo "==Validating task output ending 2==" && '
-    'echo "$s" | grep -A 5 "Job finished (status: SUCCEEDED)" | '
-    'grep "Job ID:" && '
-    'echo "$s" | grep -A 1 "Job ID:" | grep "Useful Commands"')
-
-
-# ---------- A minimal task ----------
-def test_minimal(generic_cloud: str):
-    name = _get_cluster_name()
-    test = Test(
-        'minimal',
-        [
-            f'unset SKYPILOT_DEBUG; s=$(sky launch -y -c {name} --cloud {generic_cloud} tests/test_yamls/minimal.yaml) && {_VALIDATE_LAUNCH_OUTPUT}',
-            # Output validation done.
-            f'sky logs {name} 1 --status',
-            f'sky logs {name} --status | grep "Job 1: SUCCEEDED"',  # Equivalent.
-            # Test launch output again on existing cluster
-            f'unset SKYPILOT_DEBUG; s=$(sky launch -y -c {name} --cloud {generic_cloud} tests/test_yamls/minimal.yaml) && {_VALIDATE_LAUNCH_OUTPUT}',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} --status | grep "Job 2: SUCCEEDED"',  # Equivalent.
-            # Check the logs downloading
-            f'log_path=$(sky logs {name} 1 --sync-down | grep "Job 1 logs:" | sed -E "s/^.*Job 1 logs: (.*)\\x1b\\[0m/\\1/g") && echo "$log_path" && test -f $log_path/run.log',
-            # Ensure the raylet process has the correct file descriptor limit.
-            f'sky exec {name} "prlimit -n --pid=\$(pgrep -f \'raylet/raylet --raylet_socket_name\') | grep \'"\'1048576 1048576\'"\'"',
-            f'sky logs {name} 3 --status',  # Ensure the job succeeded.
-            # Install jq for the next test.
-            f'sky exec {name} \'sudo apt-get update && sudo apt-get install -y jq\'',
-            # Check the cluster info
-            f'sky exec {name} \'echo "$SKYPILOT_CLUSTER_INFO" | jq .cluster_name | grep {name}\'',
-            f'sky logs {name} 5 --status',  # Ensure the job succeeded.
-            f'sky exec {name} \'echo "$SKYPILOT_CLUSTER_INFO" | jq .cloud | grep -i {generic_cloud}\'',
-            f'sky logs {name} 6 --status',  # Ensure the job succeeded.
-            # Test '-c' for exec
-            f'sky exec -c {name} echo',
-            f'sky logs {name} 7 --status',
-            f'sky exec echo -c {name}',
-            f'sky logs {name} 8 --status',
-            f'sky exec -c {name} echo hi test',
-            f'sky logs {name} 9 | grep "hi test"',
-            f'sky exec {name} && exit 1 || true',
-            f'sky exec -c {name} && exit 1 || true',
-        ],
-        f'sky down -y {name}',
-        _get_timeout(generic_cloud),
-    )
-    run_one_test(test)
-
-
-# ---------- Test fast launch ----------
-def test_launch_fast(generic_cloud: str):
-    name = _get_cluster_name()
-
-    test = Test(
-        'test_launch_fast',
-        [
-            # First launch to create the cluster
-            f'unset SKYPILOT_DEBUG; s=$(sky launch -y -c {name} --cloud {generic_cloud} --fast tests/test_yamls/minimal.yaml) && {_VALIDATE_LAUNCH_OUTPUT}',
-            f'sky logs {name} 1 --status',
-
-            # Second launch to test fast launch - should not reprovision
-            f'unset SKYPILOT_DEBUG; s=$(sky launch -y -c {name} --fast tests/test_yamls/minimal.yaml) && '
-            ' echo "$s" && '
-            # Validate that cluster was not re-launched.
-            '! echo "$s" | grep -A 1 "Launching on" | grep "is up." && '
-            # Validate that setup was not re-run.
-            '! echo "$s" | grep -A 1 "Running setup on" | grep "running setup" && '
-            # Validate that the task ran and finished.
-            'echo "$s" | grep -A 1 "task run finish" | grep "Job finished (status: SUCCEEDED)"',
-            f'sky logs {name} 2 --status',
-            f'sky status -r {name} | grep UP',
-        ],
-        f'sky down -y {name}',
-        timeout=_get_timeout(generic_cloud),
-    )
-    run_one_test(test)
-
-
-# See cloud exclusion explanations in test_autostop
-@pytest.mark.no_fluidstack
-@pytest.mark.no_lambda_cloud
-@pytest.mark.no_ibm
-@pytest.mark.no_kubernetes
-def test_launch_fast_with_autostop(generic_cloud: str):
-    name = _get_cluster_name()
-    # Azure takes ~ 7m15s (435s) to autostop a VM, so here we use 600 to ensure
-    # the VM is stopped.
-    autostop_timeout = 600 if generic_cloud == 'azure' else 250
-    test = Test(
-        'test_launch_fast_with_autostop',
-        [
-            # First launch to create the cluster with a short autostop
-            f'unset SKYPILOT_DEBUG; s=$(sky launch -y -c {name} --cloud {generic_cloud} --fast -i 1 tests/test_yamls/minimal.yaml) && {_VALIDATE_LAUNCH_OUTPUT}',
-            f'sky logs {name} 1 --status',
-            f'sky status -r {name} | grep UP',
-
-            # Ensure cluster is stopped
-            _WAIT_UNTIL_CLUSTER_STATUS_CONTAINS.format(
-                cluster_name=name,
-                cluster_status=ClusterStatus.STOPPED.value,
-                timeout=autostop_timeout),
-
-            # Launch again. Do full output validation - we expect the cluster to re-launch
-            f'unset SKYPILOT_DEBUG; s=$(sky launch -y -c {name} --fast -i 1 tests/test_yamls/minimal.yaml) && {_VALIDATE_LAUNCH_OUTPUT}',
-            f'sky logs {name} 2 --status',
-            f'sky status -r {name} | grep UP',
-        ],
-        f'sky down -y {name}',
-        timeout=_get_timeout(generic_cloud) + autostop_timeout,
-    )
-    run_one_test(test)
-
-
-# ---------- Test region ----------
-@pytest.mark.aws
-def test_aws_region():
-    name = _get_cluster_name()
-    test = Test(
-        'aws_region',
-        [
-            f'sky launch -y -c {name} --region us-east-2 examples/minimal.yaml',
-            f'sky exec {name} examples/minimal.yaml',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky status --all | grep {name} | grep us-east-2',  # Ensure the region is correct.
-            f'sky exec {name} \'echo $SKYPILOT_CLUSTER_INFO | jq .region | grep us-east-2\'',
-            f'sky logs {name} 2 --status',  # Ensure the job succeeded.
-            # A user program should not access SkyPilot runtime env python by default.
-            f'sky exec {name} \'which python | grep {constants.SKY_REMOTE_PYTHON_ENV_NAME} && exit 1 || true\'',
-            f'sky logs {name} 3 --status',  # Ensure the job succeeded.
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.aws
-def test_aws_with_ssh_proxy_command():
-    name = _get_cluster_name()
-
-    with tempfile.NamedTemporaryFile(mode='w') as f:
-        f.write(
-            textwrap.dedent(f"""\
-        aws:
-            ssh_proxy_command: ssh -W %h:%p -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null jump-{name}
-        """))
-        f.flush()
-        test = Test(
-            'aws_with_ssh_proxy_command',
-            [
-                f'sky launch -y -c jump-{name} --cloud aws --cpus 2 --region us-east-1',
-                # Use jump config
-                f'export SKYPILOT_CONFIG={f.name}; '
-                f'sky launch -y -c {name} --cloud aws --cpus 2 --region us-east-1 echo hi',
-                f'sky logs {name} 1 --status',
-                f'export SKYPILOT_CONFIG={f.name}; sky exec {name} echo hi',
-                f'sky logs {name} 2 --status',
-                # Start a small job to make sure the controller is created.
-                f'sky jobs launch -n {name}-0 --cloud aws --cpus 2 --use-spot -y echo hi',
-                # Wait other tests to create the job controller first, so that
-                # the job controller is not launched with proxy command.
-                _get_cmd_wait_until_cluster_status_contains_wildcard(
-                    cluster_name_wildcard='sky-jobs-controller-*',
-                    cluster_status=ClusterStatus.UP.value,
-                    timeout=300),
-                f'export SKYPILOT_CONFIG={f.name}; sky jobs launch -n {name} --cpus 2 --cloud aws --region us-east-1 -yd echo hi',
-                _WAIT_UNTIL_MANAGED_JOB_STATUS_CONTAINS_MATCHING_JOB_NAME.
-                format(
-                    job_name=name,
-                    job_status=
-                    f'({ManagedJobStatus.SUCCEEDED.value}|{ManagedJobStatus.RUNNING.value}|{ManagedJobStatus.STARTING.value})',
-                    timeout=300),
-            ],
-            f'sky down -y {name} jump-{name}; sky jobs cancel -y -n {name}',
-        )
-        run_one_test(test)
-
-
-@pytest.mark.gcp
-def test_gcp_region_and_service_account():
-    name = _get_cluster_name()
-    test = Test(
-        'gcp_region',
-        [
-            f'sky launch -y -c {name} --region us-central1 --cloud gcp tests/test_yamls/minimal.yaml',
-            f'sky exec {name} tests/test_yamls/minimal.yaml',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky exec {name} \'curl -H "Metadata-Flavor: Google" "http://metadata.google.internal/computeMetadata/v1/instance/service-accounts/default/identity?format=standard&audience=gcp"\'',
-            f'sky logs {name} 2 --status',  # Ensure the job succeeded.
-            f'sky status --all | grep {name} | grep us-central1',  # Ensure the region is correct.
-            f'sky exec {name} \'echo $SKYPILOT_CLUSTER_INFO | jq .region | grep us-central1\'',
-            f'sky logs {name} 3 --status',  # Ensure the job succeeded.
-            # A user program should not access SkyPilot runtime env python by default.
-            f'sky exec {name} \'which python | grep {constants.SKY_REMOTE_PYTHON_ENV_NAME} && exit 1 || true\'',
-            f'sky logs {name} 4 --status',  # Ensure the job succeeded.
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.ibm
-def test_ibm_region():
-    name = _get_cluster_name()
-    region = 'eu-de'
-    test = Test(
-        'region',
-        [
-            f'sky launch -y -c {name} --cloud ibm --region {region} examples/minimal.yaml',
-            f'sky exec {name} --cloud ibm examples/minimal.yaml',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky status --all | grep {name} | grep {region}',  # Ensure the region is correct.
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.azure
-def test_azure_region():
-    name = _get_cluster_name()
-    test = Test(
-        'azure_region',
-        [
-            f'sky launch -y -c {name} --region eastus2 --cloud azure tests/test_yamls/minimal.yaml',
-            f'sky exec {name} tests/test_yamls/minimal.yaml',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky status --all | grep {name} | grep eastus2',  # Ensure the region is correct.
-            f'sky exec {name} \'echo $SKYPILOT_CLUSTER_INFO | jq .region | grep eastus2\'',
-            f'sky logs {name} 2 --status',  # Ensure the job succeeded.
-            f'sky exec {name} \'echo $SKYPILOT_CLUSTER_INFO | jq .zone | grep null\'',
-            f'sky logs {name} 3 --status',  # Ensure the job succeeded.
-            # A user program should not access SkyPilot runtime env python by default.
-            f'sky exec {name} \'which python | grep {constants.SKY_REMOTE_PYTHON_ENV_NAME} && exit 1 || true\'',
-            f'sky logs {name} 4 --status',  # Ensure the job succeeded.
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-# ---------- Test zone ----------
-@pytest.mark.aws
-def test_aws_zone():
-    name = _get_cluster_name()
-    test = Test(
-        'aws_zone',
-        [
-            f'sky launch -y -c {name} examples/minimal.yaml --zone us-east-2b',
-            f'sky exec {name} examples/minimal.yaml --zone us-east-2b',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky status --all | grep {name} | grep us-east-2b',  # Ensure the zone is correct.
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.ibm
-def test_ibm_zone():
-    name = _get_cluster_name()
-    zone = 'eu-de-2'
-    test = Test(
-        'zone',
-        [
-            f'sky launch -y -c {name} --cloud ibm examples/minimal.yaml --zone {zone}',
-            f'sky exec {name} --cloud ibm examples/minimal.yaml --zone {zone}',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky status --all | grep {name} | grep {zone}',  # Ensure the zone is correct.
-        ],
-        f'sky down -y {name} {name}-2 {name}-3',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.gcp
-def test_gcp_zone():
-    name = _get_cluster_name()
-    test = Test(
-        'gcp_zone',
-        [
-            f'sky launch -y -c {name} --zone us-central1-a --cloud gcp tests/test_yamls/minimal.yaml',
-            f'sky exec {name} --zone us-central1-a --cloud gcp tests/test_yamls/minimal.yaml',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky status --all | grep {name} | grep us-central1-a',  # Ensure the zone is correct.
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-# ---------- Test the image ----------
-@pytest.mark.aws
-def test_aws_images():
-    name = _get_cluster_name()
-    test = Test(
-        'aws_images',
-        [
-            f'sky launch -y -c {name} --image-id skypilot:gpu-ubuntu-1804 examples/minimal.yaml',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky launch -c {name} --image-id skypilot:gpu-ubuntu-2004 examples/minimal.yaml && exit 1 || true',
-            f'sky launch -y -c {name} examples/minimal.yaml',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} --status | grep "Job 2: SUCCEEDED"',  # Equivalent.
-            f'sky exec {name} \'echo $SKYPILOT_CLUSTER_INFO | jq .cloud | grep -i aws\'',
-            f'sky logs {name} 3 --status',  # Ensure the job succeeded.
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.gcp
-def test_gcp_images():
-    name = _get_cluster_name()
-    test = Test(
-        'gcp_images',
-        [
-            f'sky launch -y -c {name} --image-id skypilot:gpu-debian-10 --cloud gcp tests/test_yamls/minimal.yaml',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky launch -c {name} --image-id skypilot:cpu-debian-10 --cloud gcp tests/test_yamls/minimal.yaml && exit 1 || true',
-            f'sky launch -y -c {name} tests/test_yamls/minimal.yaml',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} --status | grep "Job 2: SUCCEEDED"',  # Equivalent.
-            f'sky exec {name} \'echo $SKYPILOT_CLUSTER_INFO | jq .cloud | grep -i gcp\'',
-            f'sky logs {name} 3 --status',  # Ensure the job succeeded.
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.azure
-def test_azure_images():
-    name = _get_cluster_name()
-    test = Test(
-        'azure_images',
-        [
-            f'sky launch -y -c {name} --image-id skypilot:gpu-ubuntu-2204 --cloud azure tests/test_yamls/minimal.yaml',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky launch -c {name} --image-id skypilot:v1-ubuntu-2004 --cloud azure tests/test_yamls/minimal.yaml && exit 1 || true',
-            f'sky launch -y -c {name} tests/test_yamls/minimal.yaml',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} --status | grep "Job 2: SUCCEEDED"',  # Equivalent.
-            f'sky exec {name} \'echo $SKYPILOT_CLUSTER_INFO | jq .cloud | grep -i azure\'',
-            f'sky logs {name} 3 --status',  # Ensure the job succeeded.
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.aws
-def test_aws_image_id_dict():
-    name = _get_cluster_name()
-    test = Test(
-        'aws_image_id_dict',
-        [
-            # Use image id dict.
-            f'sky launch -y -c {name} examples/per_region_images.yaml',
-            f'sky exec {name} examples/per_region_images.yaml',
-            f'sky exec {name} "ls ~"',
-            f'sky logs {name} 1 --status',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} 3 --status',
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.gcp
-def test_gcp_image_id_dict():
-    name = _get_cluster_name()
-    test = Test(
-        'gcp_image_id_dict',
-        [
-            # Use image id dict.
-            f'sky launch -y -c {name} tests/test_yamls/gcp_per_region_images.yaml',
-            f'sky exec {name} tests/test_yamls/gcp_per_region_images.yaml',
-            f'sky exec {name} "ls ~"',
-            f'sky logs {name} 1 --status',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} 3 --status',
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.aws
-def test_aws_image_id_dict_region():
-    name = _get_cluster_name()
-    test = Test(
-        'aws_image_id_dict_region',
-        [
-            # YAML has
-            #   image_id:
-            #       us-west-2: skypilot:gpu-ubuntu-1804
-            #       us-east-2: skypilot:gpu-ubuntu-2004
-            # Use region to filter image_id dict.
-            f'sky launch -y -c {name} --region us-east-1 examples/per_region_images.yaml && exit 1 || true',
-            f'sky status | grep {name} && exit 1 || true',  # Ensure the cluster is not created.
-            f'sky launch -y -c {name} --region us-east-2 examples/per_region_images.yaml',
-            # Should success because the image id match for the region.
-            f'sky launch -c {name} --image-id skypilot:gpu-ubuntu-2004 examples/minimal.yaml',
-            f'sky exec {name} --image-id skypilot:gpu-ubuntu-2004 examples/minimal.yaml',
-            f'sky exec {name} --image-id skypilot:gpu-ubuntu-1804 examples/minimal.yaml && exit 1 || true',
-            f'sky logs {name} 1 --status',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} 3 --status',
-            f'sky status --all | grep {name} | grep us-east-2',  # Ensure the region is correct.
-            # Ensure exec works.
-            f'sky exec {name} --region us-east-2 examples/per_region_images.yaml',
-            f'sky exec {name} examples/per_region_images.yaml',
-            f'sky exec {name} --cloud aws --region us-east-2 "ls ~"',
-            f'sky exec {name} "ls ~"',
-            f'sky logs {name} 4 --status',
-            f'sky logs {name} 5 --status',
-            f'sky logs {name} 6 --status',
-            f'sky logs {name} 7 --status',
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.gcp
-def test_gcp_image_id_dict_region():
-    name = _get_cluster_name()
-    test = Test(
-        'gcp_image_id_dict_region',
-        [
-            # Use region to filter image_id dict.
-            f'sky launch -y -c {name} --region us-east1 tests/test_yamls/gcp_per_region_images.yaml && exit 1 || true',
-            f'sky status | grep {name} && exit 1 || true',  # Ensure the cluster is not created.
-            f'sky launch -y -c {name} --region us-west3 tests/test_yamls/gcp_per_region_images.yaml',
-            # Should success because the image id match for the region.
-            f'sky launch -c {name} --cloud gcp --image-id projects/ubuntu-os-cloud/global/images/ubuntu-1804-bionic-v20230112 tests/test_yamls/minimal.yaml',
-            f'sky exec {name} --cloud gcp --image-id projects/ubuntu-os-cloud/global/images/ubuntu-1804-bionic-v20230112 tests/test_yamls/minimal.yaml',
-            f'sky exec {name} --cloud gcp --image-id skypilot:cpu-debian-10 tests/test_yamls/minimal.yaml && exit 1 || true',
-            f'sky logs {name} 1 --status',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} 3 --status',
-            f'sky status --all | grep {name} | grep us-west3',  # Ensure the region is correct.
-            # Ensure exec works.
-            f'sky exec {name} --region us-west3 tests/test_yamls/gcp_per_region_images.yaml',
-            f'sky exec {name} tests/test_yamls/gcp_per_region_images.yaml',
-            f'sky exec {name} --cloud gcp --region us-west3 "ls ~"',
-            f'sky exec {name} "ls ~"',
-            f'sky logs {name} 4 --status',
-            f'sky logs {name} 5 --status',
-            f'sky logs {name} 6 --status',
-            f'sky logs {name} 7 --status',
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.aws
-def test_aws_image_id_dict_zone():
-    name = _get_cluster_name()
-    test = Test(
-        'aws_image_id_dict_zone',
-        [
-            # YAML has
-            #   image_id:
-            #       us-west-2: skypilot:gpu-ubuntu-1804
-            #       us-east-2: skypilot:gpu-ubuntu-2004
-            # Use zone to filter image_id dict.
-            f'sky launch -y -c {name} --zone us-east-1b examples/per_region_images.yaml && exit 1 || true',
-            f'sky status | grep {name} && exit 1 || true',  # Ensure the cluster is not created.
-            f'sky launch -y -c {name} --zone us-east-2a examples/per_region_images.yaml',
-            # Should success because the image id match for the zone.
-            f'sky launch -y -c {name} --image-id skypilot:gpu-ubuntu-2004 examples/minimal.yaml',
-            f'sky exec {name} --image-id skypilot:gpu-ubuntu-2004 examples/minimal.yaml',
-            # Fail due to image id mismatch.
-            f'sky exec {name} --image-id skypilot:gpu-ubuntu-1804 examples/minimal.yaml && exit 1 || true',
-            f'sky logs {name} 1 --status',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} 3 --status',
-            f'sky status --all | grep {name} | grep us-east-2a',  # Ensure the zone is correct.
-            # Ensure exec works.
-            f'sky exec {name} --zone us-east-2a examples/per_region_images.yaml',
-            f'sky exec {name} examples/per_region_images.yaml',
-            f'sky exec {name} --cloud aws --region us-east-2 "ls ~"',
-            f'sky exec {name} "ls ~"',
-            f'sky logs {name} 4 --status',
-            f'sky logs {name} 5 --status',
-            f'sky logs {name} 6 --status',
-            f'sky logs {name} 7 --status',
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.gcp
-def test_gcp_image_id_dict_zone():
-    name = _get_cluster_name()
-    test = Test(
-        'gcp_image_id_dict_zone',
-        [
-            # Use zone to filter image_id dict.
-            f'sky launch -y -c {name} --zone us-east1-a tests/test_yamls/gcp_per_region_images.yaml && exit 1 || true',
-            f'sky status | grep {name} && exit 1 || true',  # Ensure the cluster is not created.
-            f'sky launch -y -c {name} --zone us-central1-a tests/test_yamls/gcp_per_region_images.yaml',
-            # Should success because the image id match for the zone.
-            f'sky launch -y -c {name} --cloud gcp --image-id skypilot:cpu-debian-10 tests/test_yamls/minimal.yaml',
-            f'sky exec {name} --cloud gcp --image-id skypilot:cpu-debian-10 tests/test_yamls/minimal.yaml',
-            # Fail due to image id mismatch.
-            f'sky exec {name} --cloud gcp --image-id skypilot:gpu-debian-10 tests/test_yamls/minimal.yaml && exit 1 || true',
-            f'sky logs {name} 1 --status',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} 3 --status',
-            f'sky status --all | grep {name} | grep us-central1',  # Ensure the zone is correct.
-            # Ensure exec works.
-            f'sky exec {name} --cloud gcp --zone us-central1-a tests/test_yamls/gcp_per_region_images.yaml',
-            f'sky exec {name} tests/test_yamls/gcp_per_region_images.yaml',
-            f'sky exec {name} --cloud gcp --region us-central1 "ls ~"',
-            f'sky exec {name} "ls ~"',
-            f'sky logs {name} 4 --status',
-            f'sky logs {name} 5 --status',
-            f'sky logs {name} 6 --status',
-            f'sky logs {name} 7 --status',
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.aws
-def test_clone_disk_aws():
-    name = _get_cluster_name()
-    test = Test(
-        'clone_disk_aws',
-        [
-            f'sky launch -y -c {name} --cloud aws --region us-east-2 --retry-until-up "echo hello > ~/user_file.txt"',
-            f'sky launch --clone-disk-from {name} -y -c {name}-clone && exit 1 || true',
-            f'sky stop {name} -y',
-            _WAIT_UNTIL_CLUSTER_STATUS_CONTAINS.format(
-                cluster_name=name,
-                cluster_status=ClusterStatus.STOPPED.value,
-                timeout=60),
-            # Wait for EC2 instance to be in stopped state.
-            # TODO: event based wait.
-            'sleep 60',
-            f'sky launch --clone-disk-from {name} -y -c {name}-clone --cloud aws -d --region us-east-2 "cat ~/user_file.txt | grep hello"',
-            f'sky launch --clone-disk-from {name} -y -c {name}-clone-2 --cloud aws -d --region us-east-2 "cat ~/user_file.txt | grep hello"',
-            f'sky logs {name}-clone 1 --status',
-            f'sky logs {name}-clone-2 1 --status',
-        ],
-        f'sky down -y {name} {name}-clone {name}-clone-2',
-        timeout=30 * 60,
-    )
-    run_one_test(test)
-
-
-@pytest.mark.gcp
-def test_clone_disk_gcp():
-    name = _get_cluster_name()
-    test = Test(
-        'clone_disk_gcp',
-        [
-            f'sky launch -y -c {name} --cloud gcp --zone us-east1-b --retry-until-up "echo hello > ~/user_file.txt"',
-            f'sky launch --clone-disk-from {name} -y -c {name}-clone && exit 1 || true',
-            f'sky stop {name} -y',
-            f'sky launch --clone-disk-from {name} -y -c {name}-clone --cloud gcp --zone us-central1-a "cat ~/user_file.txt | grep hello"',
-            f'sky launch --clone-disk-from {name} -y -c {name}-clone-2 --cloud gcp --zone us-east1-b "cat ~/user_file.txt | grep hello"',
-            f'sky logs {name}-clone 1 --status',
-            f'sky logs {name}-clone-2 1 --status',
-        ],
-        f'sky down -y {name} {name}-clone {name}-clone-2',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.gcp
-def test_gcp_mig():
-    name = _get_cluster_name()
-    region = 'us-central1'
-    test = Test(
-        'gcp_mig',
-        [
-            f'sky launch -y -c {name} --gpus t4 --num-nodes 2 --image-id skypilot:gpu-debian-10 --cloud gcp --region {region} tests/test_yamls/minimal.yaml',
-            f'sky logs {name} 1 --status',  # Ensure the job succeeded.
-            f'sky launch -y -c {name} tests/test_yamls/minimal.yaml',
-            f'sky logs {name} 2 --status',
-            f'sky logs {name} --status | grep "Job 2: SUCCEEDED"',  # Equivalent.
-            # Check MIG exists.
-            f'gcloud compute instance-groups managed list --format="value(name)" | grep "^sky-mig-{name}"',
-            f'sky autostop -i 0 --down -y {name}',
-            _WAIT_UNTIL_CLUSTER_IS_NOT_FOUND.format(cluster_name=name,
-                                                    timeout=120),
-            f'gcloud compute instance-templates list | grep "sky-it-{name}"',
-            # Launch again with the same region. The original instance template
-            # should be removed.
-            f'sky launch -y -c {name} --gpus L4 --num-nodes 2 --region {region} nvidia-smi',
-            f'sky logs {name} 1 | grep "L4"',
-            f'sky down -y {name}',
-            f'gcloud compute instance-templates list | grep "sky-it-{name}" && exit 1 || true',
-        ],
-        f'sky down -y {name}',
-        env={'SKYPILOT_CONFIG': 'tests/test_yamls/use_mig_config.yaml'})
-    run_one_test(test)
-
-
-@pytest.mark.gcp
-def test_gcp_force_enable_external_ips():
-    name = _get_cluster_name()
-    test_commands = [
-        f'sky launch -y -c {name} --cloud gcp --cpus 2 tests/test_yamls/minimal.yaml',
-        # Check network of vm is "default"
-        (f'gcloud compute instances list --filter=name~"{name}" --format='
-         '"value(networkInterfaces.network)" | grep "networks/default"'),
-        # Check External NAT in network access configs, corresponds to external ip
-        (f'gcloud compute instances list --filter=name~"{name}" --format='
-         '"value(networkInterfaces.accessConfigs[0].name)" | grep "External NAT"'
-        ),
-        f'sky down -y {name}',
-    ]
-    skypilot_config = 'tests/test_yamls/force_enable_external_ips_config.yaml'
-    test = Test('gcp_force_enable_external_ips',
-                test_commands,
-                f'sky down -y {name}',
-                env={'SKYPILOT_CONFIG': skypilot_config})
-    run_one_test(test)
-
-
-@pytest.mark.aws
-def test_image_no_conda():
-    name = _get_cluster_name()
-    test = Test(
-        'image_no_conda',
-        [
-            # Use image id dict.
-            f'sky launch -y -c {name} --region us-east-2 examples/per_region_images.yaml',
-            f'sky logs {name} 1 --status',
-            f'sky stop {name} -y',
-            f'sky start {name} -y',
-            f'sky exec {name} examples/per_region_images.yaml',
-            f'sky logs {name} 2 --status',
-        ],
-        f'sky down -y {name}',
-    )
-    run_one_test(test)
-
-
-@pytest.mark.no_fluidstack  # FluidStack does not support stopping instances in SkyPilot implementation
-@pytest.mark.no_kubernetes  # Kubernetes does not support stopping instances
-def test_custom_default_conda_env(generic_cloud: str):
-    name = _get_cluster_name()
-    test = Test('custom_default_conda_env', [
-        f'sky launch -c {name} -y --cloud {generic_cloud} tests/test_yamls/test_custom_default_conda_env.yaml',
-        f'sky status -r {name} | grep "UP"',
-        f'sky logs {name} 1 --status',
-        f'sky logs {name} 1 --no-follow | grep -E "myenv\\s+\\*"',
-        f'sky exec {name} tests/test_yamls/test_custom_default_conda_env.yaml',
-        f'sky logs {name} 2 --status',
-        f'sky autostop -y -i 0 {name}',
-        _WAIT_UNTIL_CLUSTER_STATUS_CONTAINS.format(
-            cluster_name=name,
-            cluster_status=ClusterStatus.STOPPED.value,
-            timeout=80),
-        f'sky start -y {name}',
-        f'sky logs {name} 2 --no-follow | grep -E "myenv\\s+\\*"',
-        f'sky exec {name} tests/test_yamls/test_custom_default_conda_env.yaml',
-        f'sky logs {name} 3 --status',
-    ], f'sky down -y {name}')
-    run_one_test(test)
 
 
 # ------------ Test stale job ------------
